Disable runtime migration for bags list
